Move hooks to file file
